Working on super_learner
Adding err conditions for if the y_variable clashes with any of the learner names.

Requiring learners to be named.

Renaming regression_formula to plural to match usage.

Splitting out a verbose_output mode for being able to interrogate the cv_rmse of the model predictions in the fitting process.

Improving the ability for using different formulas for different models.

TODO:  Should the formulas be matched to the learners via indexing, naming, or either/both?